requirements.txt
numpy
# torch>=1.1
numba
tensorboardX
easydict
pyyaml
scikit-image
tqdm
kornia
torchvision
